requirements.txt
attrs==19.1.0
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
numpy==1.16.2
Pillow==5.4.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
